# apps/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_orkai-dev_orkai_apps_api_internal_apidocs.AuthResult:
    properties:
      access_token:
        type: string
      expires_at:
        type: string
      oauth_2fa_challenge:
        type: string
      refresh_token:
        type: string
      requires_2fa:
        type: boolean
      user:
        $ref: '#/definitions/github_com_orkai-dev_orkai_apps_api_internal_model.User'
    type: object
...
  github_com_orkai-dev_orkai_apps_api_internal_apidocs.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  github_com_orkai-dev_orkai_apps_api_internal_apidocs.ProblemDetail:
    properties:
      detail:
        type: string
      errors:
        items:
          $ref: '#/definitions/github_com_orkai-dev_orkai_apps_api_internal_apierr.FieldError'
        type: array
      instance:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
...
  github_com_orkai-dev_orkai_apps_api_internal_apierr.FieldError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
...
  github_com_orkai-dev_orkai_apps_api_internal_model.Role:
    enum:
    - admin
    - member
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleMember
...
  github_com_orkai-dev_orkai_apps_api_internal_model.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      display_name:
        type: string
      email:
        description: uniqueness enforced via partial index idx_users_email_active
          (active rows only)
        type: string
      first_name:
        type: string
      id:
        type: string
      is_super_admin:
        type: boolean
      last_name:
        type: string
      org_id:
        type: string
      role:
        $ref: '#/definitions/github_com_orkai-dev_orkai_apps_api_internal_model.Role'
      two_fa_enabled:
        type: boolean
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: REST control plane for orka'i self-hosted PaaS
  title: Orkai API
  version: "1.0"
paths:
...
  /team/invitations:
    get:
      description: Requires admin role.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_orkai-dev_orkai_apps_api_internal_apidocs.ProblemDetail'
      security:
      - BearerAuth: []
      summary: List pending invitations
      tags:
      - team
    post:
      consumes:
      - application/json
      description: Requires admin role.
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_orkai-dev_orkai_apps_api_internal_apidocs.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Invite organization member
      tags:
      - team
  /team/invitations/{id}:
    delete:
      description: Requires admin role.
      parameters:
      - description: Invitation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_orkai-dev_orkai_apps_api_internal_apidocs.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Cancel invitation
      tags:
      - team
  /team/invitations/accept:
    post:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_orkai-dev_orkai_apps_api_internal_apidocs.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_orkai-dev_orkai_apps_api_internal_apidocs.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Accept team invitation
      tags:
      - team
  /team/invitations/accept-public:
    post:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_orkai-dev_orkai_apps_api_internal_apidocs.AuthResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_orkai-dev_orkai_apps_api_internal_apidocs.ProblemDetail'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_orkai-dev_orkai_apps_api_internal_apidocs.ProblemDetail'
      summary: Accept invitation with registration
      tags:
      - team
  /team/invitations/info:
    get:
      parameters:
      - description: Invitation token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_orkai-dev_orkai_apps_api_internal_apidocs.ProblemDetail'
      summary: Get invitation info by token
      tags:
      - team
  /team/members:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_orkai-dev_orkai_apps_api_internal_apidocs.ProblemDetail'
      security:
      - BearerAuth: []
      summary: List organization members
      tags:
      - team
  /team/members/{id}:
    delete:
      description: Requires admin role.
      parameters:
      - description: Member user ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_orkai-dev_orkai_apps_api_internal_apidocs.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Remove organization member
      tags:
      - team
  /team/members/{id}/role:
    patch:
      consumes:
      - application/json
      description: Requires admin role.
      parameters:
      - description: Member user ID
        in: path
        name: id
        required: true
        type: string
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_orkai-dev_orkai_apps_api_internal_apidocs.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_orkai-dev_orkai_apps_api_internal_apidocs.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Update member role
      tags:
      - team
  /teams:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_orkai-dev_orkai_apps_api_internal_apidocs.ProblemDetail'
      security:
      - BearerAuth: []
      summary: List teams
      tags:
      - team
    post:
      consumes:
      - application/json
      description: Requires admin role.
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_orkai-dev_orkai_apps_api_internal_apidocs.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Create team
      tags:
      - team
  /teams/{id}:
    delete:
      description: Requires admin role.
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_orkai-dev_orkai_apps_api_internal_apidocs.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Delete team
      tags:
      - team
  /teams/{id}/members:
    get:
      description: Requires admin role.
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_orkai-dev_orkai_apps_api_internal_apidocs.ProblemDetail'
      security:
      - BearerAuth: []
      summary: List team members
      tags:
      - team
    post:
      consumes:
      - application/json
      description: Requires admin role.
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_orkai-dev_orkai_apps_api_internal_apidocs.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_orkai-dev_orkai_apps_api_internal_apidocs.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Add team member
      tags:
      - team
  /teams/{id}/members/{userId}:
    delete:
      description: Requires admin role.
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_orkai-dev_orkai_apps_api_internal_apidocs.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Remove team member
      tags:
      - team
...
securityDefinitions:
  BearerAuth:
    description: JWT access token or API key (ork_…)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"